Create StringConverter.java
7. Accepts string as input and performs 3 separate functions
a. converts the string to proper case and returns the modified string
b. counts and returns number of special characters in the string
c. sorts the string in descending order based on ASCII values and returns the sorted string
finally the main function prints the results using the returned values from these 3 functions